requirements.txt
torch
torchaudio
whisper
pydub
scipy
phonemizer
numpy
librosa
TTS
flask
flask-cors
python-telegram-bot
